Setup compose schema
relocate types

defined specs and added types

Move compose validation to sdk

stop progress

add safeCompose
