+++ src/releaseFiles/manifest/manifest.schema.json
