+++ src/releaseFiles/setupWizard.ts/setup-wizard.schema.json
