+++ src/utils/versions/increaseFromApmVersion.ts
@@ -1,8 +1,9 @@
 import { updateComposeImageTags } from "../compose";
 import { getNextVersionFromApm } from "./getNextVersionFromApm";
-import { AllowedFormats, ReleaseFileType, ReleaseType } from "../../types";
+import { ReleaseType } from "../../types";
 import { writeReleaseFile } from "../../releaseFiles/writeReleaseFile";
 import { readReleaseFile } from "../../releaseFiles/readReleaseFile";
+import { ReleaseFileType, AllowedFormats } from "../../releaseFiles/types";
 
 export async function increaseFromApmVersion({
   type,
