+++ test/utils/format.test.ts
@@ -1,7 +1,7 @@
 import { expect } from "chai";
-import { Manifest } from "../../src/types";
 import { GitHead } from "../../src/utils/git";
 import { prettyPinataPinName } from "../../src/utils/format";
+import { Manifest } from "../../src/releaseFiles/manifest/types";
 
 describe("utils / format", () => {
   describe("prettyPinataPinName", () => {
