+++ test/validation/compose/good-docker-compose.yml
@@ -1,4 +1,4 @@
-version: "3.4"
+version: "3.5"
 services:
   beacon-chain:
     image: "beacon-chain.lighthouse-prater.dnp.dappnode.eth:0.1.0"
@@ -8,9 +8,11 @@ services:
         UPSTREAM_VERSION: v2.2.0
     volumes:
       - "beacon-data:/root/.lighthouse"
+      - "/host/path:/root/.lighthouse"
     ports:
       - 9000/tcp
       - 9000/udp
+    pid: "service:erigon"
     restart: unless-stopped
     environment:
       BEACON_API_PORT: 3500
