+++ test/validation/compose/validateCompose.test.ts
@@ -1,16 +1,18 @@
 import { expect } from "chai";
 import { readReleaseFile } from "../../../src/releaseFiles/readReleaseFile";
+import { ReleaseFileType } from "../../../src/releaseFiles/types";
 import { validateSchema } from "../../../src/releaseFiles/validateSchema";
-import { ReleaseFileType } from "../../../src/types";
 
-describe("validation / Compose", () => {
+describe.only("validation / Compose", () => {
   const composeDir = "test/validation/compose";
 
-  it("Should read and validate a valid docker-compose file", () => {
+  it.only("Should read and validate a valid docker-compose file", () => {
     const compose = readReleaseFile(ReleaseFileType.compose, {
       dir: composeDir,
       releaseFileName: "good-docker-compose.yml"
     });
+
+    console.log(compose.releaseFile.services);
 
     validateSchema({
       type: ReleaseFileType.compose,
